# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Vant API
  description: |
    Vant is a persistent memory system for AI agents. 
    
    ## Spec Alignment
    - Embeddings endpoint is OpenAI-compatible (`/v1/embeddings`)
    - All endpoints gated behind JWT/API key auth
    - QoS enforced via rate limiting
    
    ## Authentication
    Bearer token in Authorization header:
    ```
    Authorization: Bearer <token>
    ```
  version: 0.8.7
  contact:
    name: Vant Team
    url: https://vant.creadev.org
servers:
  - url: https://api.vant.creadev.org
    description: Production
  - url: http://localhost:3457
    description: Local development
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: integer
    BrainFile:
      type: object
      properties:
        name:
          type: string
        content:
          type: string
        source:
          type: string
          enum: [public, private, remote]
        updated:
          type: string
          format: date-time
    BrainList:
      type: object
      properties:
        brains:
          type: array
          items:
            $ref: '#/components/schemas/BrainFile'
...
paths:
...
  /api/v1/brain:
    get:
      tags:
        - Brain
      summary: List all brains
      operationId: listBrains
      responses:
        '200':
          description: List of brains
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrainList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Brain
      summary: Create/update brain file
      operationId: saveBrain
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                content:
                  type: string
                source:
                  type: string
                  enum: [public, private]
      responses:
        '200':
          description: Brain saved
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'